create analyzer.rs
